Move functions -> utils file
