+++ packages/stores/src/queries-external/filtered-pools/filtered-pools.ts
@@ -7,13 +7,8 @@ import { ObservableQueryPool } from "../../queries/pools/pool";
 import { IPoolGetter } from "../../queries/pools/types";
 import { IMPERATOR_HISTORICAL_DATA_BASEURL } from "..";
 import { ObservableQueryExternalBase } from "../base";
-import {
-  Filters,
-  objToQueryParams,
-  Pagination,
-  FilteredPools,
-  makePoolRawFromFilteredPool,
-} from "./types";
+import { Filters, objToQueryParams, Pagination, FilteredPools } from "./types";
+import { makePoolRawFromFilteredPool } from "./utils";
 
 /** TEMPORARY: use imperator query to fetch filtered, sorted pools */
 export class ObservableQueryFilteredPools
